+++ test/cram/lint/meta.t
@@ -16,7 +16,22 @@ Meta workitems are not checked in the database:
   > "#1115","General okra maintenance","Draft","","","","Maintenance - internal tooling","","pillar/ecosystem,team/internal-tooling",""
   > EOF
 
-  $ cat > weekly.md << EOF
+  $ cat > valid.md << EOF
+  > # Last week
+  > 
+  > - Off
+  >   - @eng1 (1 day)
+  > 
+  > - Hack
+  >   - @eng1 (1 day)
+  >   - xxx
+  > 
+  > - Misc
+  >   - @eng1 (3 days)
+  >   - xxx
+  > EOF
+
+  $ cat > invalid.md << EOF
   > # Last week
   > 
   > - Leave
@@ -31,5 +46,48 @@ Meta workitems are not checked in the database:
   >   - xxx
   > EOF
 
-  $ okra lint -e -C admin/ weekly.md
-  [OK]: weekly.md
+If the report is out of a admin/weekly repository, we accept the meta categories:
+
+  $ okra lint -e -C admin/ valid.md
+  [OK]: valid.md
+
+Before week 24, the categories didn't exist:
+
+  $ mkdir -p admin/weekly/2024/23
+  $ cp valid.md admin/weekly/2024/23
+  $ okra lint -e -C admin/ admin/weekly/2024/23/valid.md
+  File "admin/weekly/2024/23/valid.md", line 3:
+  Error: In objective "Off":
+         No work items found. This may indicate an unreported parsing error. Remove the objective if it is without work.
+  File "admin/weekly/2024/23/valid.md", line 3:
+  Error: In objective "Off":
+         No ID found. Objectives should be in the format "This is an objective (#123)", where 123 is the objective issue ID. For objectives that don't have an ID yet, use "New KR" and for work without an objective use "No KR".
+  File "admin/weekly/2024/23/valid.md", line 6:
+  Error: In objective "Hack":
+         No ID found. Objectives should be in the format "This is an objective (#123)", where 123 is the objective issue ID. For objectives that don't have an ID yet, use "New KR" and for work without an objective use "No KR".
+  File "admin/weekly/2024/23/valid.md", line 10:
+  Error: In objective "Misc":
+         No ID found. Objectives should be in the format "This is an objective (#123)", where 123 is the objective issue ID. For objectives that don't have an ID yet, use "New KR" and for work without an objective use "No KR".
+  [1]
+
+Starting from week 24, they are supported:
+
+  $ mkdir -p admin/weekly/2024/24
+  $ cp valid.md admin/weekly/2024/24
+  $ okra lint -e -C admin/ admin/weekly/2024/24/valid.md
+  [OK]: admin/weekly/2024/24/valid.md
+
+The previous meta categories are not supported anymore:
+
+  $ cp invalid.md admin/weekly/2024/24
+  $ okra lint -e -C admin/ admin/weekly/2024/24/invalid.md
+  File "admin/weekly/2024/24/invalid.md", line 3:
+  Error: In objective "Leave":
+         No work items found. This may indicate an unreported parsing error. Remove the objective if it is without work.
+  File "admin/weekly/2024/24/invalid.md", line 3:
+  Error: In objective "Leave":
+         No ID found. Objectives should be in the format "This is an objective (#123)", where 123 is the objective issue ID. For objectives that don't have an ID yet, use "New KR" and for work without an objective use "No KR".
+  File "admin/weekly/2024/24/invalid.md", line 10:
+  Error: In objective "Off-objective":
+         No ID found. Objectives should be in the format "This is an objective (#123)", where 123 is the objective issue ID. For objectives that don't have an ID yet, use "New KR" and for work without an objective use "No KR".
+  [1]