+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### Changed
 
+- Lint: using standardized categories before Week 24 (June 10, 2024) raises an error (#273, @gpetiot). Only Off, Hack and Misc are now valid.
 - Lint: using a work-item instead of an objective raises an error starting from Week 24 (June 10, 2024) (#270, @gpetiot)
 - Lint: do not fail if an objective is used for the wrong quarter (#272, @gpetiot)